+++ CHANGELOG.md
@@ -1,4 +1,28 @@
-# Changelog
+## 0.9.10
+
+* Remove `isInitialRoute` from full screen page route
+
+## 0.9.9
+
+* Changed wakelock plugin from `flutter_screen` to `wakelock` due to lack of maintenance of `flutter_screen`. 
+
+## 0.9.8+1
+  * Require latest flutter stable version
+
+## 0.9.8
+
+  * Hero Widget is no longer used (thanks @localpcguy)
+  * Tap to hide controls (thanks @bostrot)
+  * Replay on play when video is finished (thanks @VictorUvarov)
+
+## 0.9.7
+
+  * Errors are properly handled. You can provide the Widget to display when an error occurs by providing an `errorBuilder` function to the `ChewieController` constructor.
+  * Add ability to override the fullscreen page builder. Allows folks to customize that functionality!
+
+## 0.9.6
+
+  * Update to work with `video_player: ">=0.7.0 <0.11.0"`
 
 ## 0.9.5